server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = "~=3.11"
dependencies = [
    "flask",
    "langchain",
    "langchain-google-genai",
    "pypdf",
    "faiss-cpu",
    "tiktoken",
    "python-dotenv",
    "flask-cors",
    "gunicorn"
]